Nelson

+++ src/main/java/nelsonssoares/ecomproductsapi/domain/repository/SubCategoriaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubCategoriaRepository extends JpaRepository<SubCategoria, Long> {
+public interface SubCategoriaRepository extends JpaRepository<SubCategoria, Integer> {
 }
